wwwroot/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gcloud==0.18.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==4.1.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.8
Pyrebase4==4.4.1
python-jwt==2.0.1
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
urllib3==1.25.11
Werkzeug==1.0.1
